% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_by_idx.R
\name{compare_by_idx}
\alias{compare_by_idx}
\title{compare_by_idx}
\usage{
compare_by_idx(
  tree_df,
  idx1,
  idx2,
  chrom1 = NULL,
  chrom2 = NULL,
  treetype = "trees",
  use.edge.length = TRUE,
  space = 0,
  length.line = 1,
  gap = 2,
  type = "phylogram",
  rotate = FALSE,
  col = par("fg"),
  lwd = par("lwd"),
  lty = par("lty"),
  show.tip.label = TRUE,
  font = 3
)
}
\arguments{
\item{tree_df}{tree \code{data.frame} produced by
\code{get_topologies} [mandatory]}

\item{idx1}{index (chromosome index if chrom1 is set)
for first tree [mandatory]}

\item{idx2}{index (chromosome index if chrom2 is set)
for second tree [mandatory]}

\item{chrom1}{select chromosome for first tree [optional]}

\item{chrom2}{select chromosome for second tree [optional]}

\item{treetype}{select tree to compare (trees, rooted_trees, topoloy_trees)}

\item{use.edge.length}{a logical indicating whether the branch lengths
should be used to plot the trees [FALSE]}

\item{space}{a positive value that specifies the distance between the two
trees [0]}

\item{length.line}{a positive value that specifies the length of the
horizontal line associated to each taxa [1]}

\item{gap}{a value specifying the distance between the tips of the phylogeny
and the lines [2]}

\item{type}{a character string specifying the type of phylogeny to be drawn
(phylogram, cladogram)}

\item{rotate}{a logical indicating whether the nodes of the phylogeny can
be rotated by clicking [FALSE]}

\item{col}{a character vector indicating the color to be used for the links;
recycled as necessary}

\item{lwd}{id. for the width}

\item{lty}{d. for the line type}

\item{show.tip.label}{a logical indicating whether to show the tip labels on
the phylogeny [TRUE]}

\item{font}{an integer specifying the type of font for the
labels: 1 (plain text),
2 (bold),
3 (italic, the default),
or 4 (bold italic)}
}
\description{
This function plots two trees face to face given two indices.
}
\examples{
trees <- ape::rmtree(N = 100, n = 4, rooted = FALSE)
topologies <- TreeSwitchR::get_topologies(trees)
TreeSwitchR::compare_by_idx(
    tree_df = topologies[["tree_df"]],
    idx1 = 1,
    idx2 = 2)
}
\author{
Kristian K Ullrich
}
