% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos.R
\name{create_pos}
\alias{create_pos}
\title{create_pos}
\usage{
create_pos(n)
}
\arguments{
\item{n}{number of trees [mandatory]}
}
\description{
This function creates a simple position matrix of lnegth n.
}
\examples{
pos <- TreeSwitchR::create_pos(n = 10)
pos
}
\author{
Kristian K Ullrich
}
