% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_to_face.R
\name{face_to_face}
\alias{face_to_face}
\title{face_to_face}
\usage{
face_to_face(
  xtree,
  ytree,
  layout = "rectangular",
  link_color = par("fg"),
  tip_color = par("fg"),
  link_space = 0.2,
  ...
)
}
\arguments{
\item{xtree}{xtree}

\item{ytree}{ytree}

\item{layout}{layout}

\item{link_color}{link color}

\item{tip_color}{tip color}

\item{link_space}{link space}
}
\description{
This function plots two trees face to face.
}
\examples{
trees <- ape::rmtree(N = 100, n = 4, rooted = FALSE)
topologies <- TreeSwitchR::get_topologies(trees)
xytrees <- TreeSwitchR::get_by_idx(
    tree_df = topologies[["tree_df"]],
    idx1 = 1,
    idx2 = 2)
TreeSwitchR::face_to_face(
    xtree = xytrees[["xtree"]],
    ytree = xytrees[["ytree"]])
}
\author{
Kristian K Ullrich
}
