% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_by_pos.R
\name{get_by_pos}
\alias{get_by_pos}
\title{get_by_pos}
\usage{
get_by_pos(
  tree_df,
  chrom1,
  chromStart1,
  chromEnd1,
  chrom2,
  chromStart2,
  chromEnd2,
  treetype = "trees"
)
}
\arguments{
\item{tree_df}{tree \code{data.frame} produced by
\code{get_topologies} [mandatory]}

\item{chrom1}{select chromosome for first tree [mandatory]}

\item{chromStart1}{select chromosome start for first tree [mandatory]}

\item{chromEnd1}{select chromosome end for first tree [mandatory]}

\item{chrom2}{select chromosome for second tree [mandatory]}

\item{chromStart2}{select chromosome start for second tree [mandatory]}

\item{chromEnd2}{select chromosome end for second tree [mandatory]}

\item{treetype}{select tree to compare (trees, rooted_trees, topoloy_trees)}
}
\description{
This function returns two trees given two positions.
}
\examples{
trees <- ape::rmtree(N = 100, n = 4, rooted = FALSE)
topologies <- TreeSwitchR::get_topologies(trees)
TreeSwitchR::get_by_pos(
    tree_df = topologies[["tree_df"]],
    chrom1 = "chrom",
    chromStart1 = 1,
    chromEnd1 = 2,
    chrom2 = "chrom",
    chromStart2 = 5,
    chromEnd2 = 6)
}
\author{
Kristian K Ullrich
}
