% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_not_consecutive.R
\name{get_not_consecutive}
\alias{get_not_consecutive}
\title{get_not_consecutive}
\usage{
get_not_consecutive(x)
}
\arguments{
\item{x}{vector to find non consecutive positions [mandatory]}
}
\description{
This function returns the index of non consecutive positions.
}
\examples{
x <- c("a", "a", "b", "b", "c", "d")
TreeSwitchR::get_not_consecutive(x)
}
\author{
Kristian K Ullrich
}
