% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topologies.R
\name{get_topologies}
\alias{get_topologies}
\title{get_topologies}
\usage{
get_topologies(trees, pos = NULL, outgroup = NULL, sort = TRUE)
}
\arguments{
\item{trees}{phylogenetic trees [mandatory]}

\item{pos}{tree positions [optional]}

\item{outgroup}{[optional]}

\item{sort}{specify if trees should be sorted by tree position first [TRUE]}
}
\description{
This function creates tree topologies from phylogenetic trees.
The output is a \code{list} which contains:
1. a \code{data.frame} tree_df with the columns:
original trees, rooted trees, topo trees, topology,
chrom, chromStart, chromEnd
2. trees
3. rooted_trees
4. topo_trees
5. topo_trees_table_sorted
6. topos
7. pos
}
\examples{
trees <- ape::rmtree(N = 100, n = 4, rooted = FALSE)
data("pos", package = "TreeSwitchR")
topologies <- TreeSwitchR::get_topologies(trees, pos)
tree_df <- topologies[["tree_df"]]
tree_gr <- topologies[["tree_gr"]]
}
\seealso{
\code{\link[Gviz]{GeneRegionTrack}}

\code{\link[Gviz]{plotTracks}}
}
\author{
Kristian K Ullrich
}
