% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gtf_txdb.R
\name{make_gtf_txdb}
\alias{make_gtf_txdb}
\title{make_gtf_txdb}
\usage{
make_gtf_txdb(gtffile)
}
\arguments{
\item{gtffile}{GFF3/GTF file [mandatory]}
}
\description{
This function makes a \code{TxDb} object from a GFF3/GTF file.
}
\examples{
# GTF
gtffile <- tempfile(pattern = "gtf.")
download.file(url = paste0("https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/",
"000/006/765/",
"GCF_000006765.1_ASM676v1/GCF_000006765.1_ASM676v1_genomic.gtf.gz"),
destfile = gtffile)
gtf_txdb <- TreeSwitchR::make_gtf_txdb(gtffile)
# GFF3
gff3file <- tempfile(pattern = "gff3.")
download.file(url = paste0("https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/",
"000/006/765/",
"GCF_000006765.1_ASM676v1/GCF_000006765.1_ASM676v1_genomic.gff.gz"),
destfile = gff3file)
gff3_txdb <- TreeSwitchR::make_gtf_txdb(gff3file)
}
\seealso{
\code{\link[txdbmaker]{makeTxDbFromGFF}}
}
\author{
Kristian K Ullrich
}
