% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_tree_pos.R
\name{order_tree_pos}
\alias{order_tree_pos}
\title{order_tree_pos}
\usage{
order_tree_pos(trees, pos)
}
\arguments{
\item{trees}{phylogenetic trees [mandatory]}

\item{pos}{[mandatory]}
}
\description{
This function orders trees by tree position (chrom, chromStart).
The output is a \code{list} which contains:
1. ordered trees
2. ordered positions
}
\examples{
data("pos", package="TreeSwitchR")
trees <- ape::rmtree(N = 100, n = 4, rooted = FALSE)
ordered <- TreeSwitchR::order_tree_pos(trees, pos)
}
\author{
Kristian K Ullrich
}
