% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatures.R
\name{plotFeatures}
\alias{plotFeatures}
\title{plotFeatures}
\usage{
plotFeatures(txdb, chromosome, from, to)
}
\arguments{
\item{txdb}{\code{TxDB} object [mandatory]}

\item{chromosome}{chromosome [mandatory]}

\item{from}{region start position in bp [mandatory]}

\item{to}{region end position in bp [mandatory]}
}
\description{
This function plots annotated Features of a \code{TxDB} object
for a given chromosome region.
}
\examples{
# GTF
gtffile <- tempfile(pattern = "gtf.")
download.file(url = paste0("https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/",
"000/006/765/",
"GCF_000006765.1_ASM676v1/GCF_000006765.1_ASM676v1_genomic.gtf.gz"),
destfile = gtffile)
gtf_txdb <- TreeSwitchR::make_gtf_txdb(gtffile)
TreeSwitchR::plotFeatures(txdb = gtf_txdb,
    chromosome = "NC_002516.2",
    from = 1,
    to = 100000)
}
\seealso{
\code{\link[Gviz]{GeneRegionTrack}}

\code{\link[Gviz]{plotTracks}}
}
\author{
Kristian K Ullrich
}
