% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeSummary.R
\name{plotTreeSummary}
\alias{plotTreeSummary}
\title{plotTreeSummary}
\usage{
plotTreeSummary(
  tree_df,
  chrom = NULL,
  chromStart = NULL,
  chromEnd = NULL,
  chromPad = 0,
  colorBy = "rank",
  chromSplitColor = "red",
  plotType = "point",
  pointSize = 0.5,
  pointShape = 15
)
}
\arguments{
\item{tree_df}{tree \code{data.frame} produced by
\code{get_topologies} [mandatory]}

\item{chrom}{select chromosome to plot summary [optional]}

\item{chromStart}{select chromosome start to plot [optional]}

\item{chromEnd}{select chromosome end to plot [optional]}

\item{chromPad}{distance between chromosomes [0]}

\item{colorBy}{color tree topologies either by rank [default] or by counts}

\item{chromSplitColor}{color of chrom split [red]}

\item{plotType}{plot type either by point [default] or lollipop}

\item{pointSize}{point size [0.5]}

\item{pointShape}{point shape [15]}
}
\description{
This function plots tree topologies distribution along
chromosomes.
}
\examples{
trees <- ape::rmtree(N = 100, n = 4, rooted = FALSE)
data("pos", package = "TreeSwitchR")
topologies <- TreeSwitchR::get_topologies(trees, pos)
tree_df <- topologies[["tree_df"]]
p1 <- TreeSwitchR::plotTreeSummary(
    tree_df,
    colorBy = "rank")
p2 <- TreeSwitchR::plotTreeSummary(
    tree_df,
    colorBy = "counts")
p3 <- TreeSwitchR::plotTreeSummary(
    tree_df,
    colorBy = "rank",
    chrom = "chr1")
p4 <- TreeSwitchR::plotTreeSummary(
    tree_df,
    colorBy = "rank",
    chrom = "chr2")
plotly::subplot(
    p1[["fig"]],
    p2[["fig"]],
    p3[["fig"]],
    p4[["fig"]])

}
\author{
Kristian K Ullrich
}
