% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos-data.R
\docType{data}
\name{pos-data}
\alias{pos-data}
\alias{pos}
\title{pos-data}
\format{
an object of class \code{tibble}
}
\usage{
data(pos)
}
\description{
Example position matrix.
}
\examples{
data("pos", package="TreeSwitchR")
}
\keyword{datasets}
