% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtf.R
\name{read_gtf}
\alias{read_gtf}
\title{read_gtf}
\usage{
read_gtf(gtffile)
}
\arguments{
\item{gtffile}{GFF3/GTF file [mandatory]}
}
\description{
This function reads a GFF3/GTF file.
}
\examples{
# GTF
gtffile <- tempfile(pattern = "gtf.")
download.file(file = paste0("https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/",
"000/006/765/",
"GCF_000006765.1_ASM676v1/GCF_000006765.1_ASM676v1_genomic.gtf.gz"),
destfile = gtffile)
gtf <- TreeSwitchR::read_gtf(gtffile)
# GFF3
gff3file <- tempfile(pattern = "gff3.")
download.file(file = paste0("https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/",
"000/006/765/",
"GCF_000006765.1_ASM676v1/GCF_000006765.1_ASM676v1_genomic.gff.gz"),
destfile = gff3file)
gff3_txdb <- TreeSwitchR::read_gtf(gff3file)
}
\seealso{
\code{\link[rtracklayer]{import}}
}
\author{
Kristian K Ullrich
}
