% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_posfile.R
\name{read_posfile}
\alias{read_posfile}
\title{read_posfile}
\usage{
read_posfile(posfile, header = TRUE, sort = TRUE)
}
\arguments{
\item{posfile}{tree position file [mandatory]}

\item{header}{header (chrom<tab>chromStart<tab>chromEnd) exists [TRUE]}

\item{sort}{sort by chrom and chromStart [TRUE]}
}
\description{
This function reads a chromosome position file.
Each line corresponds to one tree.
chrom<tab>chromStart<tab>chromEnd
}
\examples{
data("pos", package="TreeSwitchR")
posfile <- tempfile()
write.table(
    pos,
    sep = "\t",
    quote = FALSE,
    col.names = TRUE,
    row.names = FALSE,
    file = posfile)
pos <- TreeSwitchR::read_posfile(posfile)
pos
}
\seealso{
\code{\link[readr]{read_tsv}}
}
\author{
Kristian K Ullrich
}
