% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_treefile.R
\name{read_treefile}
\alias{read_treefile}
\title{read_treefile}
\usage{
read_treefile(treefile)
}
\arguments{
\item{treefile}{treefile [mandatory]}
}
\description{
This function reads newick trees from a file.
}
\examples{
trees <- ape::rmtree(N = 100, n = 3, rooted = FALSE)
treefile <- tempfile()
ape::write.tree(trees, file = treefile)
trees <- TreeSwitchR::read_treefile(treefile)
trees
}
\seealso{
\code{\link[ape]{read.tree}}
}
\author{
Kristian K Ullrich
}
