% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSwitchR-package.R
\docType{package}
\name{TreeSwitchR-package}
\alias{TreeSwitchR-package}
\title{TreeSwitchR: TreeSwitchR visualize phylogenetic trees along chromosomes in R}
\description{
TreeSwitchR visualize phylogenetic trees along chromosomes in R. Given a tree file and the corresponding genomic positions, trees are visualized and compared to their corresponding neighbors. Tree topology switches as well as their underlying genomic features are highlighted.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kullrich/TreeSwitchR}
  \item Report bugs at \url{https://github.com/kullrich/TreeSwitchR/issues}
}

}
\author{
\strong{Maintainer}: Kristian K Ullrich \email{ullrich@evolbio.mpg.de} (\href{https://orcid.org/0000-0003-4308-9626}{ORCID})

}
\keyword{internal}
