% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSwitchR.R
\name{TreeSwitchR}
\alias{TreeSwitchR}
\title{TreeSwitchR}
\usage{
TreeSwitchR(
  tree_df = NULL,
  treefile = NULL,
  posfile = NULL,
  gtffile = NULL,
  return.app = TRUE
)
}
\arguments{
\item{return.app}{Optional boolean indicating whether a Shiny app should be
returned [TRUE]
If \code{FALSE}, a named list of app elements (ui and server) will be
returned instead. Useful for deploying as a standalone shiny app.}
}
\description{
TreeSwitchR
}
\examples{
library(TreeSwitchR)
# Create app with no data loaded.
app <- TreeSwitchR()
if(interactive()) {
    shiny::runApp(app)
}
}
\author{
Kristian K Ullrich
}
