% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopologies.R
\name{plotTopologies}
\alias{plotTopologies}
\title{plotTopologies}
\usage{
plotTopologies(
  tree_df,
  nrow = NULL,
  ncol = NULL,
  title = "Tree Topologies",
  tip_color = par("fg"),
  ...
)
}
\arguments{
\item{tree_df}{tree \code{data.frame} produced by
\code{get_topologies} [mandatory]}

\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{title}{title}

\item{tip_color}{tip color}
}
\description{
This function plots topologies given a \code{tree_df}
}
\examples{
trees <- ape::rmtree(N = 100, n = 4, rooted = FALSE)
topologies <- TreeSwitchR::get_topologies(trees)
plotTopologies(
    tree_df = topologies[["tree_df"]])
}
\author{
Kristian K Ullrich
}
