% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeSummary.R
\name{plotTreeSummary}
\alias{plotTreeSummary}
\title{plotTreeSummary}
\usage{
plotTreeSummary(tree_df, chrom = NULL, chromPad = 1)
}
\arguments{
\item{tree_df}{tree \code{data.frame} produced by
\code{get_topologies} [mandatory]}

\item{chrom}{select chromosome to plot summary [optional]}

\item{chromPad}{distance between chromosomes [optional]}
}
\description{
This function plots tree topologies distribution along
chromsomes.
}
\examples{
trees <- ape::rmtree(N = 100, n = 4, rooted = FALSE)
topologies <- TreeSwitchR::get_topologies(trees)
TreeSwitchR::plotTreeSummary(topologies[["tree_df"]])
}
\author{
Kristian K Ullrich
}
